# services/api/docs/swagger-spec.yaml
openapi: 3.0.0
paths:
  /api/posts/create:
    post:
      operationId: PostsController_createPost
      summary: Responsible for post creation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostDto"
      responses:
        "200":
          description: As a response function gets success message
        "201":
          description: ""
          content:
            application/json:
              schema:
                type: object
      tags:
        - Posts
  /api/user/sign-up:
    post:
      operationId: UserController_signUp
      summary: Responsible for user account creation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUpDto"
      responses:
        "201":
          description: As a response function gets 2 tokens - refresh and access
      tags: &a1
        - User
  /api/user/sign-in:
    post:
      operationId: UserController_signIn
      summary: Responsible for user login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInDto"
      responses:
        "201":
          description: As a response function gets success message
      tags: *a1
  "/api/user/account-confirmation/{confirmationHash}":
    get:
      operationId: UserController_confirmAccount
      summary: Confirm user registration
      parameters:
        - name: confirmationHash
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
        "201":
          description: As a response function gets success message
      tags: *a1
  /api/user/logout:
    post:
      operationId: UserController_logout
      parameters: []
      responses:
        "201":
          description: ""
      tags: *a1
info:
  title: Cryptotalks - API Gateway
  description: Documentation of Cryptotalks' API Gateway.
  version: 0.0.1
  contact: {}
tags: []
servers: []
components:
  schemas:
    Post:
      type: object
      properties:
        id:
          type: uuidv4
          nullable: false
          default: uuidv4
          description: Unique Id of the record
        title:
          type: string
          nullable: false
          description: Title of the post
        slug:
          type: string
          nullable: false
          description: Slug of the post
        content:
          nullable: false
          description: Content of the post
          type: array
          items:
            type: string
        userId:
          type: uuiv4
          nullable: false
          description: User Id
        createdAt:
          type: date
          nullable: false
          description: Record creation date
        updatedAt:
          type: string
          nullable: false
          description: Record update date
      required:
        - id
        - title
        - slug
        - content
        - userId
        - createdAt
        - updatedAt
    CreatePostDto:
      type: object
      properties:
        title:
          type: string
          nullable: false
        content:
          nullable: false
          type: array
          items:
            type: string
      required:
        - title
        - content
    User:
      type: object
      properties:
        id:
          type: uuidv4
          nullable: false
          default: uuidv4
          description: Unique Id of the record
        email:
          type: string
          nullable: false
          description: User email
        password:
          type: string
          nullable: false
          description: User hashed password
        accountConfirm:
          type: boolean
          nullable: false
          description: Checks if user confirmed account by clicking into link in email
            after registration
          default: false
        firstName:
          type: string
          nullable: true
          description: User first name
        lastName:
          type: string
          nullable: true
          description: User last name
        username:
          type: string
          nullable: false
          description: User username
        twitter:
          type: string
          nullable: true
          description: User Twitter account
        linkedIn:
          type: string
          nullable: true
          description: User LinkedIn account
        personalWebsite:
          type: string
          nullable: true
          description: User personal website
        title:
          type: string
          nullable: true
          description: User account title
        bio:
          type: string
          nullable: true
          description: User biography
        tac:
          type: boolean
          nullable: false
          description: User accepted terms and conditions
        publicEmail:
          type: boolean
          nullable: true
          description: If user wants to set their email as public
        createdAt:
          type: date
          nullable: false
          description: Record creation date
        updatedAt:
          type: string
          nullable: false
          description: Record update date
      required:
        - id
        - email
        - password
        - accountConfirm
        - firstName
        - lastName
        - username
        - twitter
        - linkedIn
        - personalWebsite
        - title
        - bio
        - tac
        - publicEmail
        - createdAt
        - updatedAt
    SignUpDto:
      type: object
      properties:
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
          nullable: false
        password:
          type: string
          nullable: false
        username:
          type: string
          nullable: false
        twitter:
          type: string
          nullable: true
        linkedIn:
          type: string
          nullable: true
        personalWebsite:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        tac:
          type: boolean
          nullable: false
        publicEmail:
          type: boolean
          nullable: true
      required:
        - email
        - password
        - username
        - tac
    TokenPayloadDto:
      type: object
      properties:
        id:
          type: string
          nullable: false
        type:
          type: access | refresh
          nullable: false
        userId:
          type: string
          nullable: false
      required:
        - id
        - type
        - userId
    UpdateTokensDto:
      type: object
      properties:
        userId:
          type: string
          nullable: false
        email:
          type: string
          nullable: false
      required:
        - userId
        - email
    RefreshTokenDto:
      type: object
      properties:
        userId:
          type: string
          nullable: false
        tokenId:
          type: string
          nullable: false
      required:
        - userId
        - tokenId
    AccessTokenDto:
      type: object
      properties:
        userId:
          type: string
          nullable: false
        email:
          type: string
          nullable: false
      required:
        - userId
        - email
    Session:
      type: object
      properties:
        id:
          type: uuidv4
          nullable: false
          default: uuidv4
          description: Unique Id of the record
        tokenId:
          type: uuidv4
          nullable: false
          description: Id of the token
        userId:
          type: uuidv4
          nullable: false
          description: User id
        createdAt:
          type: date
          nullable: false
          description: Record creation date
        updatedAt:
          type: string
          nullable: false
          description: Record update date
      required:
        - id
        - tokenId
        - userId
        - createdAt
        - updatedAt
    SignInDto:
      type: object
      properties:
        email:
          type: string
          nullable: false
        password:
          type: string
          nullable: false
      required:
        - email
        - password
    AccountConfirmationDto:
      type: object
      properties:
        confirmationHash:
          type: string
          nullable: false
      required:
        - confirmationHash
    ConfirmationHash:
      type: object
      properties:
        id:
          type: string
          nullable: false
          default: uuidv4
          description: Unique Id of the record
        confirmationHash:
          type: string
          nullable: false
          description: Confirmation hash
        confirmed:
          type: boolean
          nullable: false
          description: If hash has been confirmed
        userId:
          type: uuidv4
          nullable: false
          description: User Id
        createdAt:
          type: date
          nullable: false
          description: Record creation date
        updatedAt:
          type: string
          nullable: false
          description: Record update date
      required:
        - id
        - confirmationHash
        - confirmed
        - userId
        - createdAt
        - updatedAt